refact(Menu): renames component to header name
