+++ src/app/(portal)/_components/header/Header.types.ts
