+++ src/app/(portal)/layout.tsx
@@ -1,5 +1,5 @@
 import Footer from "./_components/footer/Footer.component";
-import Menu from "./_components/menu/Menu.component";
+import Menu from "./_components/header/Header.component";
 
 function Layout({ children }: { children: React.ReactNode }) {
   return (
